+++ docs/search/files_5.js
@@ -3,5 +3,6 @@ var searchData=
   ['objects_2ehpp_0',['objects.hpp',['../objects_8hpp.html',1,'']]],
   ['opengl3dengine_2ehpp_1',['OpenGL3DEngine.hpp',['../_open_g_l3_d_engine_8hpp.html',1,'']]],
   ['openglengine_2ehpp_2',['OpenGLEngine.hpp',['../_open_g_l_engine_8hpp.html',1,'']]],
-  ['openglutils_2ehpp_3',['OpenGLUtils.hpp',['../_open_g_l_utils_8hpp.html',1,'']]]
+  ['openglutils_2ehpp_3',['OpenGLUtils.hpp',['../_open_g_l_utils_8hpp.html',1,'']]],
+  ['orthographiccamera_2ehpp_4',['OrthographicCamera.hpp',['../_orthographic_camera_8hpp.html',1,'']]]
 ];