+++ docs/search/files_6.js
@@ -1,7 +1,4 @@
 var searchData=
 [
-  ['shaderprogram_2ehpp_0',['ShaderProgram.hpp',['../_shader_program_8hpp.html',1,'']]],
-  ['shaderprogrampipeline_2ehpp_1',['ShaderProgramPipeline.hpp',['../_shader_program_pipeline_8hpp.html',1,'']]],
-  ['shaderutils_2ehpp_2',['ShaderUtils.hpp',['../_shader_utils_8hpp.html',1,'']]],
-  ['simpleshader_2ehpp_3',['SimpleShader.hpp',['../_simple_shader_8hpp.html',1,'']]]
+  ['perspectivecamera_2ehpp_0',['PerspectiveCamera.hpp',['../_perspective_camera_8hpp.html',1,'']]]
 ];