+++ docs/search/files_8.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['uniformbufferobject_2ehpp_0',['UniformBufferObject.hpp',['../_uniform_buffer_object_8hpp.html',1,'']]]
+  ['teapot_2ehpp_0',['teapot.hpp',['../teapot_8hpp.html',1,'']]],
+  ['textureutils_2ehpp_1',['TextureUtils.hpp',['../_texture_utils_8hpp.html',1,'']]]
 ];